Reorganisation of demo files and algorithmic change in CLAClassifier.js (only for qt_interactive)
Running the QuickTest demo one observes that there is no prediction from
Sunday to Monday, because the memory object is resetted on Sunday. This
seems to be due to the fact that the correct prediction is classified
incorrectly by the CLAClassifier. This can be fixed by a small change in
CLAClassifier.js,.